Dialog de confirmação no Angular para remoção de curso

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -7,17 +7,16 @@ import { catchError } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ConfirmationDialogComponent } from '../../components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
   templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.scss']
+  styleUrls: ['./courses.component.scss'],
 })
 export class CoursesComponent implements OnInit {
-
   courses$: Observable<Course[]> | null = null;
   displayedColumns = ['name', 'category', 'actions'];
-
 
   constructor(
     private coursesService: CoursesService,
@@ -30,45 +29,49 @@ export class CoursesComponent implements OnInit {
   }
 
   onAdd() {
-    this.router.navigate(['new'], {relativeTo: this.route});
+    this.router.navigate(['new'], { relativeTo: this.route });
   }
 
   onError(errorMassage: string) {
     this.dialog.open(ErrorDialogComponent, {
-      data: errorMassage
+      data: errorMassage,
     });
   }
 
   onEdit(course: Course) {
-    this.router.navigate(['edit', course._id], {relativeTo: this.route});
+    this.router.navigate(['edit', course._id], { relativeTo: this.route });
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-          this.refresh()
-          this.snackBar.open('Cusro removido com sucesso!', 'X', {
-            duration: 5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'
-          });
-      },
-      () => this.onError('Erro ao tentar remover Curso!')
-    )
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja reomver esse Curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: Boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Cusro removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          () => this.onError('Erro ao tentar remover Curso!')
+        );
+      }
+    });
   }
 
   refresh() {
-    this.courses$ = this.coursesService.listAll()
-    .pipe(
-      catchError(error => {
-        this.onError("Erro ao carregar Cursos.")
-        return of([])
+    this.courses$ = this.coursesService.listAll().pipe(
+      catchError((error) => {
+        this.onError('Erro ao carregar Cursos.');
+        return of([]);
       })
     );
   }
 
-
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
